C-择时类/另类ETF交易策略：日内动量/requirements.txt
backtrader==1.9.78.123
empyrical==0.5.5
empyrical_reloaded==0.5.10
loguru==0.7.2
matplotlib==3.8.4
numpy==1.23.5
pandas==2.2.2
pyfolio_reloaded==0.9.7
tqdm==4.66.4
